+++ .gitignore
@@ -89,6 +89,4 @@ fastlane/test_output
 # https://github.com/johnno1962/injectionforxcode
 
 iOSInjectionProject/
-UseBy/.DS_Store
-MyPlayground.playground/
 .DS_Store
