+++ UseBy/Components/AppName.swift
@@ -37,17 +37,3 @@ final class AppName: UILabel {
         super.init(coder: decoder)
     }
 }
-
-extension UILabel {
-  func addCharacterSpacing(kernValue: Double = 1.15) {
-    if let labelText = text, labelText.count > 0 {
-      let attributedString = NSMutableAttributedString(string: labelText)
-        attributedString.addAttribute(
-            NSAttributedString.Key.kern,
-            value: kernValue,
-            range: NSRange(location: 0, length: attributedString.length - 1)
-        )
-      attributedText = attributedString
-    }
-  }
-}
